build(Gradle): Move advisors to individual plugin projects
