--- advisor/src/main/resources/META-INF/services/org.ossreviewtoolkit.advisor.AdviceProviderFactory
@@ -1,5 +1,0 @@
-org.ossreviewtoolkit.advisor.advisors.GitHubDefects$Factory
-org.ossreviewtoolkit.advisor.advisors.NexusIq$Factory
-org.ossreviewtoolkit.advisor.advisors.OssIndex$Factory
-org.ossreviewtoolkit.advisor.advisors.Osv$Factory
-org.ossreviewtoolkit.advisor.advisors.VulnerableCode$Factory
