+++ plugins/advisors/nexus-iq/build.gradle.kts
@@ -23,12 +23,10 @@ plugins {
 }
 
 dependencies {
+    api(project(":advisor"))
     api(project(":model"))
-    api(project(":utils:common-utils"))
 
+    implementation(project(":clients:nexus-iq-client"))
+    implementation(project(":utils:common-utils"))
     implementation(project(":utils:ort-utils"))
-
-    implementation(libs.kotlinxCoroutines)
-
-    testImplementation(libs.mockk)
 }
