+++ plugins/advisors/nexus-iq/src/main/resources/META-INF/services/org.ossreviewtoolkit.advisor.AdviceProviderFactory
@@ -1,0 +1,1 @@
+org.ossreviewtoolkit.plugins.advisors.nexusiq.NexusIq$Factory
