+++ plugins/advisors/oss-index/src/test/assets/__files/response_components.json
