+++ plugins/advisors/osv/src/main/resources/META-INF/services/org.ossreviewtoolkit.advisor.AdviceProviderFactory
@@ -1,0 +1,1 @@
+org.ossreviewtoolkit.plugins.advisors.osv.Osv$Factory
