+++ plugins/advisors/vulnerable-code/src/test/assets/__files/README.md
@@ -1,0 +1,4 @@
+This directory contains test files for tests using the [WireMock](http://wiremock.org/) framework.
+
+It is a convention of the test framework that the files to be served by the mock server must be located below a
+directory named `__files` (also see [this issue](https://github.com/wiremock/wiremock/issues/1163)).
