+++ plugins/advisors/vulnerable-code/src/test/assets/__files/response_invalid_uri.json
