+++ plugins/advisors/vulnerable-code/src/test/assets/__files/response_junit.json
