+++ plugins/advisors/vulnerable-code/src/test/assets/__files/response_log4j.json
