+++ plugins/advisors/vulnerable-code/src/test/assets/__files/response_packages_no_vulnerabilities.json
