+++ plugins/advisors/vulnerable-code/src/test/assets/ort-analyzer-result.yml
